fix #159 - feature added: copy links into other linkgroups

+++ classes/class.JTL-Shop.Link.php
@@ -422,7 +422,9 @@ class Link extends MainModel
 
             if (is_array($oLink_arr) && count($oLink_arr) > 0) {
                 foreach ($oLink_arr as &$oLink) {
+                    $kLinkgruppe = $oLink->kLinkgruppe;
                     $oLink = new self($oLink->kLink, null, true);
+                    $oLink->kLinkgruppe = (int)$kLinkgruppe;
                 }
             }
 
@@ -513,27 +515,35 @@ class Link extends MainModel
 
     /**
      * @param bool $bSub
+     * @param int  $kLinkgruppe
      * @return int
      */
-    public function delete($bSub = true)
+    public function delete($bSub = true, $kLinkgruppe = null)
     {
         $nRows = 0;
         if ($this->kLink > 0) {
-            $nRows = Shop::DB()->query("DELETE FROM tlink WHERE kLink = " . $this->getLink(), 3);
-            Shop::DB()->delete('tlinksprache', 'kLink', $this->getLink());
-            Shop::DB()->query("DELETE FROM tseo WHERE kKey = " . $this->getLink() . " AND cKey = 'kLink'", 4);
+            if (!empty($kLinkgruppe)) {
+                $nRows = Shop::DB()->query("DELETE FROM tlink WHERE kLink = " . $this->getLink() . " AND kLinkgruppe = " . $kLinkgruppe, 3);
+            } else {
+                $nRows = Shop::DB()->query("DELETE FROM tlink WHERE kLink = " . $this->getLink(), 3);
+            }
+            $nLinkAnz = Shop::DB()->query("SELECT * FROM tlink WHERE kLink = " . $this->getLink(), 2);
+            if (count($nLinkAnz) === 0) {
+                Shop::DB()->delete('tlinksprache', 'kLink', $this->getLink());
+                Shop::DB()->query("DELETE FROM tseo WHERE kKey = " . $this->getLink() . " AND cKey = 'kLink'", 4);
 
-            $cDir = PFAD_ROOT . PFAD_BILDER . PFAD_LINKBILDER . $this->getLink();
-            if (is_dir($cDir) && $this->getLink() > 0) {
-                if (delDirRecursively($cDir)) {
-                    rmdir($cDir);
+                $cDir = PFAD_ROOT . PFAD_BILDER . PFAD_LINKBILDER . $this->getLink();
+                if (is_dir($cDir) && $this->getLink() > 0) {
+                    if (delDirRecursively($cDir)) {
+                        rmdir($cDir);
+                    }
                 }
             }
 
             if ($bSub) {
                 if (isset($this->oSub_arr) && count($this->oSub_arr) > 0) {
                     foreach ($this->oSub_arr as $oSub) {
-                        $oSub->delete();
+                        $oSub->delete(true, $kLinkgruppe);
                     }
                 }
             }
